+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/whitespace/parenpad/InputParenPadForWhitespace.java
@@ -2,9 +2,9 @@
 // Test case file for FOR_ITERATION and whitespace.
 // Created: 2003
 ////////////////////////////////////////////////////////////////////////////////
-package com.puppycrawl.tools.checkstyle.checks.whitespace;
+package com.puppycrawl.tools.checkstyle.checks.whitespace.parenpad;
 
-class InputForWhitespace
+class InputParenPadForWhitespace
 {
     void method1()
     {
