+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/whitespace/parenpad/InputParenPadLambda.java
@@ -1,6 +1,6 @@
-package com.puppycrawl.tools.checkstyle.checks.whitespace;
+package com.puppycrawl.tools.checkstyle.checks.whitespace.parenpad;
 
-public class InputParenPadLambdaOnly {
+class InputParenPadLambda {
     {
         java.util.function.Consumer a = (o) -> { o.toString(); }; // ok
 
@@ -16,9 +16,9 @@ public class InputParenPadLambdaOnly {
 
         java.util.stream.Stream.of().forEach(( Object o ) -> o.toString()); // 2 violations
 
-        java.util.stream.Stream.of().forEach(o -> o.toString( )); // ok
+        java.util.stream.Stream.of().forEach(o -> o.toString( )); // 2 violations
     }
 
-    void someMethod( String param ) { // ok
+    void someMethod( String param ) { // 2 violations
     }
 }
