+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/whitespace/parenpad/InputParenPadLeftRightAndNoSpace.java
@@ -1,10 +1,10 @@
-package com.puppycrawl.tools.checkstyle.checks.whitespace;
+package com.puppycrawl.tools.checkstyle.checks.whitespace.parenpad;
 
 import java.io.IOException;
 import java.io.StringWriter;
 import java.io.Writer;
 
-public class InputParenPad
+public class InputParenPadLeftRightAndNoSpace
 {
     class ParenPadNoSpace  {
         ParenPadNoSpace() {
